UI (#5)
* Add files via upload

Home/Login/Registration Screens

* Add files via upload

So I am adding in the AlertAcademyAccountSettings.html codes in the UI file and the style.css that is connected to the account setting so I wont forget that I created it during the weekend.

* Update style.css

I was just adding more coding for the style.css that is connected to the account setting. But I am still trying to add in more later on. - Angelu

* Update style.css

Still trying to add in more coding for the style.css for the account setting

* Update style.css

need to fix the text size since I accidentally did 7em instead of .7em

* Update style.css

* Update style.css

* Update style.css

* Update AlertAcademyAccountSettings.html

I was adding in the gender and pronoun since I know fit we publish this and test the product it would work well.

* Update AlertAcademyAccountSettings.html

I was changing the gender section into a radio button input and fixing up the pronoun section and using the radio button as well.

* Update style.css

* Update style.css

* Update style.css

* Update and rename CSS - AlertAcademy.css to CSS - AlertAcademy1.css

* Update and rename HTML - AlertAcademy.html to HTML - AlertAcademy1.html

---------

Co-authored-by: LarryAnders <149206985+LarryAnders@users.noreply.github.com>
Co-authored-by: angeluJD <156857396+angeluJD@users.noreply.github.com>